requirements.txt
# flask==3.1.1
numpy==2.3.1
geopandas==1.1.1
dotenv==0.9.9
openai==0.28
rasterio==1.4.3
earthengine_api==1.6.0
folium==0.20.0
ipython==9.4.0
# flask_cors==6.0.
fastapi==0.116.1
python-multipart==0.0.20
uvicorn==0.35.0
pydantic-settings==2.10.1
pytest==8.4.1